fix: RowContainer::estimateRowSize 无符号除法致排序输出批退化为 1 行/批
  usedSize（int64）在 Windows 上因 freeBytes 按可寻址余量计而为负时，
  与 uint64_t 的 numRows_ 相除发生无符号除法，负值被包装成约 1e14 的
  巨商，绕过既有的 rowSize<=0 → fixedRowSize_ 回退；OrderBy/TopN/
  TopNRowNumber/HashAgg/StreamingAgg/Window 的输出批大小因此系统性
  钳到 1 行/批（全排序回退路径实测 200 万批、200 万次 allocation）。

  修复：除法前将 numRows_ 显式转为 int64_t，恢复 Windows 回退语义。

+++ velox/exec/RowContainer.cpp
@@ -1057,7 +1057,11 @@ std::optional<int64_t> RowContainer::estimateRowSize() const {
   int64_t usedSize = rows_.allocatedBytes() - freeBytes +
       stringAllocator_->retainedSize() - stringAllocator_->freeSpace() -
       rowPointers_.capacity() * sizeof(char*);
-  int64_t rowSize = usedSize / numRows_;
+  // 'numRows_' is unsigned; a negative 'usedSize' (Windows allocation
+  // accounting, see below) must not wrap into a huge positive quotient via
+  // unsigned division, or the <=0 fallback below never fires and callers
+  // collapse their output batch size to 1 row.
+  int64_t rowSize = usedSize / static_cast<int64_t>(numRows_);
   // On Windows, allocation overhead accounting may differ from Linux causing
   // usedSize to be zero or negative for small row counts. Fall back to the
   // fixed row size which is always a valid lower bound.
